- moved all the tests from suite_context_test.go that do NOT rely on "features" and it's massive set of steps into run_test - removed some duplicate tests from example_subtests_test.go
