+++ Makefile
@@ -21,11 +21,12 @@ check-go-version:
 	fi
 
 test: check-go-version
-	@echo "running all tests"
-	@go fmt ./...
-	@go run honnef.co/go/tools/cmd/staticcheck@v0.4.7 github.com/cucumber/godog
-	@go run honnef.co/go/tools/cmd/staticcheck@v0.4.7 github.com/cucumber/godog/cmd/godog
+	@echo checks
+	go fmt ./...
+	go run honnef.co/go/tools/cmd/staticcheck@v0.5.1 github.com/cucumber/godog
+	go run honnef.co/go/tools/cmd/staticcheck@v0.5.1 github.com/cucumber/godog/cmd/godog
 	go vet ./...
+	@echo "running all tests"
 	go test -race ./...
 	go run ./cmd/godog -f progress -c 4
