--- example_subtests_test.go
@@ -1,45 +1,0 @@
-package godog_test
-
-import (
-	"testing"
-
-	"github.com/cucumber/godog"
-)
-
-func ExampleTestSuite_Run_subtests() {
-	var t *testing.T // Comes from your test function, e.g. func TestFeatures(t *testing.T).
-
-	suite := godog.TestSuite{
-		ScenarioInitializer: func(s *godog.ScenarioContext) {
-			// Add step definitions here.
-		},
-		Options: &godog.Options{
-			Format:   "pretty",
-			Paths:    []string{"features"},
-			TestingT: t, // Testing instance that will run subtests.
-		},
-	}
-
-	if suite.Run() != 0 {
-		t.Fatal("non-zero status returned, failed to run feature tests")
-	}
-}
-
-func TestFeatures(t *testing.T) {
-	suite := godog.TestSuite{
-		ScenarioInitializer: func(s *godog.ScenarioContext) {
-			godog.InitializeScenario(s)
-
-			// Add step definitions here.
-		},
-		Options: &godog.Options{
-			Format:   "pretty",
-			Paths:    []string{"features"},
-			TestingT: t, // Testing instance that will run subtests.
-		},
-	}
-
-	if suite.Run() != 0 {
-		t.Fatal("non-zero status returned, failed to run feature tests")
-	}
-}
